+++ app/mcp/solid_apm/resources/application_resource.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+module SolidApm
+  module Mcp
+    class ApplicationResource < FastMcp::Resource
+    end
+  end
+end
