+++ app/mcp/solid_apm/tools/application_tool.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+module SolidApm
+  module Mcp
+    class ApplicationTool < FastMcp::Tool
+    end
+  end
+end
