--- app/mcp_resources/application_mcp_resource.rb
@@ -1,6 +1,0 @@
-# frozen_string_literal: true
-
-module SolidApm
-  class ApplicationMcpResource < FastMcp::Resource
-  end
-end
